:truck: Moves Endpoints out of core
