+++ src/index.ts
@@ -1,9 +1,14 @@
+export { refreshToken } from './auth/refreshToken';
+export { tokensFromCode } from './auth/tokensFromCode';
 export { SPOTIFY_URL } from './constants';
-export { getCurrentUser } from './core/endpoints/users/getCurrentUser';
 export { setToken } from './core/setToken';
 export { spotifyApiClient } from './core/spotifyApiClient';
+export { getCurrentUser } from './endpoints/users/getCurrentUser';
 export { deepFreeze } from './utils/deepFreeze';
+export { isBrowser, isNode } from './utils/isBrowserOrNode';
 export { spotifyFetch } from './utils/spotifyFetch';
+export { toBase64 } from './utils/toBase64';
+export type { SpotifyTokens } from './auth/tokensFromCode';
 export type {
   PersistentApiProperties,
   SpotifyApiClient,
